no negative scores for losing HQ
Outputting -1 for a normal loss without a bot crashing causes
the viewer to output '-' (no score) and brutaltester to detect bot
crashes.

This change makes stdout brutaltester compatible, by altering the
referee output to adjust it's end game condition losing score to 0,
freeing -1 to be used to represent an invalid input or a timeout.